Update for local installer

+++ filter-devkit-mavenizer/src/main/java/com/vordel/mavenizer/InstallScriptGenerator.java
@@ -9,6 +9,8 @@ import java.nio.file.StandardCopyOption;
 import java.util.HashSet;
 import java.util.Set;
 
+import org.apache.maven.model.Model;
+import org.apache.maven.model.Parent;
 import org.apache.maven.repository.internal.ArtifactDescriptorUtils;
 import org.eclipse.aether.RepositorySystem;
 import org.eclipse.aether.RepositorySystemSession;
@@ -30,6 +32,7 @@ import org.eclipse.aether.util.artifact.JavaScopes;
 import com.vordel.mavenizer.dist.AetherUtils;
 import com.vordel.mavenizer.dist.GatewayRepoSys;
 import com.vordel.mavenizer.dist.MavenizerRepository;
+import com.vordel.mavenizer.maps.MavenMetaMap;
 
 public class InstallScriptGenerator {
 	public static void usage(String format, Object... args) {
@@ -79,6 +82,7 @@ public class InstallScriptGenerator {
 	public static void install(GatewayRepoSys sys, Set<Artifact> artifacts) {
 		File root = new File("install").getAbsoluteFile();
 		Set<Artifact> installed = new HashSet<Artifact>();
+		Set<Artifact> poms = new HashSet<Artifact>();
 
 		if (!root.exists()) {
 			root.mkdir();
@@ -86,7 +90,6 @@ public class InstallScriptGenerator {
 
 		RepositorySystemSession session = sys.getOfflineSession();
 		RepositorySystem system = sys.getSystem();
-
 		StringBuilder builder = new StringBuilder();
 
 		builder.append("#!/bin/sh\n");
@@ -102,72 +105,82 @@ public class InstallScriptGenerator {
 
 				dependencyRequest.setRoot(collectResult.getRoot());
 
-				resolveMissing(sys.getSession(), system, collectResult.getRoot());
+				if (resolveMissing(sys.getSession(), system, collectResult.getRoot())) {
+					DependencyResult dependencyResult = system.resolveDependencies(session, dependencyRequest);
 
-				DependencyResult dependencyResult = system.resolveDependencies(session, dependencyRequest);
+					for (ArtifactResult artifactResult : dependencyResult.getArtifactResults()) {
+						if (artifactResult.isResolved()) {
+							Artifact item = artifactResult.getArtifact();
+							Artifact pom = null;
 
-				for (ArtifactResult artifactResult : dependencyResult.getArtifactResults()) {
-					if (artifactResult.isResolved()) {
-						Artifact item = artifactResult.getArtifact();
-						Artifact pom = null;
-
-						if (installed.add(item)) {
-							try {
-								if (!"pom".equals(item.getExtension())) {
-									try {
-										ArtifactResult pomResult = system.resolveArtifact(session, new ArtifactRequest().setArtifact(ArtifactDescriptorUtils.toPomArtifact(item)));
-
-										if (pomResult.isResolved()) {
-											pom = pomResult.getArtifact();
-										}
-									} catch (ArtifactResolutionException e) {
-										/* ignore this */
-									}
-								}
-
-								if ("jar".equals(item.getExtension())) {
-									String classifier = item.getClassifier();
-
-									if (!"sources".equals(classifier)) {
+							if (installed.add(item)) {
+								try {
+									if (!"pom".equals(item.getExtension())) {
 										try {
-											ArtifactResult sourcesResult = system.resolveArtifact(session, new ArtifactRequest().setArtifact(new DefaultArtifact(item.getGroupId(), item.getArtifactId(), "sources", item.getExtension(), item.getVersion())));
+											ArtifactResult pomResult = system.resolveArtifact(session, new ArtifactRequest().setArtifact(ArtifactDescriptorUtils.toPomArtifact(item)));
 
-											if (sourcesResult.isResolved()) {
-												appendArtifact(root, builder, sourcesResult.getArtifact(), null);
+											if (pomResult.isResolved()) {
+												pom = pomResult.getArtifact();
+
+												if (pom != null) {
+													poms.add(pom);
+												}
 											}
 										} catch (ArtifactResolutionException e) {
 											/* ignore this */
 										}
+									} else {
+										poms.add(item);
 									}
 
-									if (!"javadoc".equals(classifier)) {
-										try {
-											ArtifactResult sourcesResult = system.resolveArtifact(session, new ArtifactRequest().setArtifact(new DefaultArtifact(item.getGroupId(), item.getArtifactId(), "javadoc", item.getExtension(), item.getVersion())));
+									if ("jar".equals(item.getExtension())) {
+										String classifier = item.getClassifier();
 
-											if (sourcesResult.isResolved()) {
-												appendArtifact(root, builder, sourcesResult.getArtifact(), null);
+										if (!"sources".equals(classifier)) {
+											try {
+												ArtifactResult sourcesResult = system.resolveArtifact(session, new ArtifactRequest().setArtifact(new DefaultArtifact(item.getGroupId(), item.getArtifactId(), "sources", item.getExtension(), item.getVersion())));
+
+												if (sourcesResult.isResolved()) {
+													appendArtifact(root, builder, sourcesResult.getArtifact(), null);
+												}
+											} catch (ArtifactResolutionException e) {
+												/* ignore this */
 											}
-										} catch (ArtifactResolutionException e) {
-											/* ignore this */
+										}
+
+										if (!"javadoc".equals(classifier)) {
+											try {
+												ArtifactResult sourcesResult = system.resolveArtifact(session, new ArtifactRequest().setArtifact(new DefaultArtifact(item.getGroupId(), item.getArtifactId(), "javadoc", item.getExtension(), item.getVersion())));
+
+												if (sourcesResult.isResolved()) {
+													appendArtifact(root, builder, sourcesResult.getArtifact(), null);
+												}
+											} catch (ArtifactResolutionException e) {
+												/* ignore this */
+											}
 										}
 									}
-								}
 
-								appendArtifact(root, builder, item, pom);
-							} catch (IOException e) {
-								/* ignore this */
+									appendArtifact(root, builder, item, pom);
+								} catch (IOException e) {
+									/* ignore this */
+								}
 							}
 						}
 					}
 				}
+			}
 
-				FileWriter writer = new FileWriter(new File(root, "install.sh"));
+			for(Artifact artifact : poms) {
+				appendParent(session, system, root, builder, artifact, installed);
+			}
 
-				try {
-					writer.append(builder);
-				} finally {
-					writer.close();
-				}
+			FileWriter writer = new FileWriter(new File(root, "install.sh"));
+
+			try {
+				writer.append(builder);
+			} finally {
+				writer.close();
 			}
 		} catch (DependencyResolutionException e) {
 			usage("got exception during dependency resolution");
@@ -177,10 +190,45 @@ public class InstallScriptGenerator {
 		} catch (IOException e) {
 			usage("got exception writing deploy script");
 		}
-
 	}
 
-	private static void resolveMissing(RepositorySystemSession session, RepositorySystem system, DependencyNode node) {
+	private static void appendParent(RepositorySystemSession session, RepositorySystem system, File root, StringBuilder builder, Artifact artifact, Set<Artifact> installed) throws IOException {
+		if ("pom".equals(artifact.getExtension())) {
+			Model model = MavenMetaMap.parseModel(MavenizerRepository.readFile(artifact.getFile()));
+
+			if (model != null) {
+				Parent parent = model.getParent();
+
+				if (parent != null) {
+					String groupId = parent.getGroupId();
+					String artifactId = parent.getArtifactId();
+					String version = parent.getVersion();
+
+					if ((groupId != null) && (artifactId != null) && (version != null)) {
+						ArtifactRequest request = new ArtifactRequest().setArtifact(new DefaultArtifact(groupId, artifactId, null, "pom", version));
+
+						try {
+							ArtifactResult result = system.resolveArtifact(session, request);
+
+
+							if (result.isResolved()) {
+								Artifact resolved = result.getArtifact();
+
+								if (installed.add(resolved)) {
+									appendArtifact(root, builder, resolved, null);
+								}
+
+								appendParent(session, system, root, builder, resolved, installed);
+							}
+						} catch (ArtifactResolutionException e) {
+						}
+					}
+				}
+			}
+		}
+	}
+
+	private static boolean resolveMissing(RepositorySystemSession session, RepositorySystem system, DependencyNode node) {
 		ArtifactRequest request = new ArtifactRequest();
 		ArtifactResult result = null;
 
@@ -206,13 +254,16 @@ public class InstallScriptGenerator {
 				AetherUtils.setRemoteRepositoriesProxy(session, request.getRepositories());
 				result = system.resolveArtifact(session, request);
 			} catch (ArtifactResolutionException e) {
-				usage("got exception resolving artifact");
 			}
 		}
 
-		for (DependencyNode child : node.getChildren()) {
-			resolveMissing(session, system, child);
+		if (result != null) {
+			for (DependencyNode child : node.getChildren()) {
+				resolveMissing(session, system, child);
+			}
 		}
+
+		return result != null;
 	}
 
 	private static Artifact copyArtifact(Artifact artifact, File root) throws IOException {
@@ -220,15 +271,17 @@ public class InstallScriptGenerator {
 		File out = new File(new File(root, MavenizerRepository.sha1(in)), in.getName()).getAbsoluteFile();
 		File parent = out.getParentFile();
 
-		if (parent.exists()) {
-			for (File file : parent.listFiles()) {
-				file.delete();
-			}
-
+		if (parent.exists() && (!parent.isDirectory())) {
 			parent.delete();
 		}
 
-		parent.mkdir();
+		if (!parent.exists()) {
+			parent.mkdir();
+		}
+
+		if (out.exists()) {
+			out.delete();
+		}
 
 		Files.copy(in.toPath(), out.toPath(), StandardCopyOption.COPY_ATTRIBUTES);
